Change bam_plp_insertion_mod to return number of bases, not length.
When we have no base modification, the return value and ins->l are
both the same thing (number of bases inserted).

When we have base modifications, the old length was string length (and
ironically ins->l was incorrectly left as the number of bases).
That's not particularly obvious, nor useful.

It's now reversed, so ins->l is the length of the string ins->s and
the return value is the number of bases.  Eg "+3AG[+o99]A" vs "+3AGA"
both now return 3.