Add some error detection to bam_next_basemod.
Having an MM tag refer to bases that are beyond the end of the
sequence now generates a warning and returns -1 instead of 0 from the
function. It's a warning because it's harmless to ignore these, but
the caller may wish to check and treat it as a hard error (as the test
harness does).

Note: this isn't trivial to do for the reverse strand, so (for now) we
only check the top strand data.

Also fixed a typo in a comment.