Fix cram_dependent_data_series when FN is used.
If we get into the primary loop of cram_decode_seq, iterating over
feature (count FN) then we have unguarded code that always checks for
sequence overlapping the reference.  To do this, seq_pos must be set.
Some data series we know how seq_pos is adjusted irrespective of
whether we decode, eg BS is always +1, but others are strings and the
only way we know how to update seq_pos is to decode them.

Hence added FN as having a dependency on SC, IN and BB.

Fixes samtools/samtools#1475