Make version.sh work on shallow git checkouts.
Use "git describe --always" to silence "No names found, cannot
describe anything." messages on shallow checkouts where the
last tagged release may not be available.  Where this is the case,
git describe --always returns the short commit hash.  This is
appended to the hard-coded version number to make something
similar to the normal git describe output.  As we don't know
how many commits there are from the last tag in this case, that
part is set to 1 for want of anything better.