Fix indexing bug with "placed" unmapped reads.
Unmapped-but-placed (having REF/POS) reads are not included in the
index.  Hence if an placed unmapped is the first record in a bin, then
it may not be returned.  Note most aligners write out mapped followed
by unmapped which does not trigger this problem.

The SAM spec states that all unmapped placed reads should be
considered as having an alignment length of 1.  While it doesn't seem
to explicitly state these must therefore be in the index, it does
imply it.  It appears that picard also indexes placed reads in this
manner.

Originally reported by John Marshall.  Fixes #1142