Fix excessive memory used by multi-threaded SAM output on long reads.
The previous code here buffered up 1000 BAM records before dispatching
a job to convert to SAM.  We now buffer up either 1000 BAM records or
(estimated) ~240KB of encoded SAM, whichever is hit first.

Note these figures were previously written in macros NB and NM, which
have been replaced with more meaningful names.

I tested it with data having an average length of 148bp (Illumina)
15949bp (ONT) and 13441 (PB CCS), using the GIAB HG002 chr1 public data
sets.  Maximum lengths were 148, 1044875 and 19288.

/usr/bin/time real time and maxresident figures, before:

               Illumina              ONT               PB
    -@8     43,128/2m43   6,040,568/1m39   1,593,932/1m15
    -@16    70,880/1m42   8,710,132/1m08   3,026,604/0m52
    -@32    70,272/1m29  11,940,036/0m55   5,722,952/0m47
    -@64   190,584/1m21  17,007,840/0m56  10,835,512/0m48

After:

               Illumina              ONT               PB
    -@16    50,208/1m36     696,276/1m09      63,496/0m57
    -@32    86,044/1m21   1,054,524/0m53     109,696/0m44
    -@64   149,024/1m24   1,616,720/0m55     195,676/0m48

The effect on memory (KB) is vast, although it's still a bit higher on
ONT.  This is probably related to the maximum lengths being used
and the reuse of BAM structs (never shrinking them) rather than the
average, so a long tail on the distribution causes memory growth.

We could address that in later updates, but this is still a huge
improvement.

Obviously as we get to very long records, we'll be dispatching very
commonly (maybe every alignment), but I don't yet know how inefficient
the threading becomes then.  Memory usage will grow as we cannot store
half an alignment, but it won't explode so fast.  We may wish to try
larger values of SAM_NBYTES, but note I tried 100KB to 1MB and this
came out fairly optimal on CPU so there was little reason to trade
more memory for CPU.