Correct writing CRAM alignment span field for unmapped data.
The specification states unmapped data must have alignment start and
alignment span both zero.  Previously span was 1 (from end-start+1).
We now adhere to the specification for 3.1 onwards, but it is left
as-is (incorrect) for 3.0.

See previous commit for implications on reading.  Outputting spec
compliant CRAM files would make older builds of htslib/samtools fail
to return unmapped data from a CRAM index query.  This is not helpful,
so in this case we feel the specification is best amended to permit
other values in the alignment span field (albeit to keep the existing
values as recommendations).