Check number syntax is valid in hts_parse_decimal()
Follow strtol() et al's conventions, setting strend to the original str
(or emitting a warning) if the input string does not form a valid numeric
constant. In particular, require it to have a non-zero number of digits.

Merge E and k/M/G handling, so that at most one can be used. In particular,
"1e+3k" is non-standard and ambiguous, and should be rejected.

Document the numeric syntax parsed and the two modes of operation (as
determined by whether  strend  is non-NULL).

Add test cases, in particular verifying that a sole "G" is rejected.