Improve coordinate parsing, particularly with tabix. (PR #1411)
Double check the file coordinates are 1-based unless tabix -0 is
specified.  We treat a 0 coordinate as a warning now.  The
documentation is also more explicit that -0 also implied a half-open
coordinate system.

Also improve region parsing.  The code supports regions such as
"x"/"x:" (all of "x"), "x:10-20" "x:-20" (up to pos 20) and "x:20-"
(position 20 onwards).  However as genome coordinates are 1-based and
internally we have 0-based, we subtract one during parsing.  The
"x:-20" is done by negative value detection, but this was also
triggered with the "x:0" region (treated as "up to 1").

Also, illegal regions such as "x:-10-20" were treated as "x:-20".
This is now a hard error.

Fixes #1409