Make hfile_s3 refresh AWS credentials on expiry
This is to make HTSlib work better with AWS IAM credentials,
which have a limited lifespan, and so may need to be refreshed.
To allow this, hfile_s3 is made to look for an unofficial
'expiry_time' entry in the AWS_SHARED_CREDENTIALS_FILE.  If
present, the file will be re-read if the current time is within
one minute of the given expiry (new credentails are available
five minutes before expiry, according to
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/iam-roles-for-amazon-ec2.html).

Currently no effort is made to understand the JSON format
emitted by the AWS security-credentials endpoint.  It's up
to the end user to reformat this into the style of the normal
'.aws/credentials' file.  An example of how this can be done
for one source of credentials on AWS is added to the manual
page.

Fixes bug where parse_ini would append to rather than replace
existing values.

Moves x-amz-security-token to the set of headers updated
via callback, as it can now change when the credentials
are updated.

Includes an implementation of the timegm() function, which
is not portable (e.g. mingw doesn't have it) but needed to convert
the expiry time to a time_t.  This is put in a separate header
so that it can be more easily reused elsewhere if we want.
Includes tests to check that details like leap years and
normalisation work properly.