Add an embed_ref=2 mode which starts with ref=N and edits on the fly.
We compute reference from seq + cigar + MD tag.  This gets used
immediately as it's part of the reference-based encoding in
process_one_read.

If we have inconsistent records where the inferred reference differs,
then we would end up with an invalid embedded reference and decode
errors.  Instead we just accept the first one as we cannot go back and
correct earlier mistakes.  It's a little tricky when the sequence may
contain "N" characters, as that's also the initial value of unassigned
reference, so we have to be careful there.

This also opens the way for handling CRAM files without any MD tags.
In that scenario we just assume there are no SNPs and create a fake MD
tag (just numbers merging match/mismatch and ^N* for deletions).  So
we can now do embedded reference on files with no reference and no
MD.  It's not quite consensus (which would be better), but that would
require a two-pass encoding strategy in place of process_one_read.  So
it's a bit poor on high-error technologies.

Note, when embedded auto-generated reference where MD tags may be
absent, the auto-decoded MD may be incorrect as the reference we have
embedded is now the sequence as-is rather than the real sequence.