Make htscodecs SIMD code build without configure
For the benefit of htslib embeds that don't want to create and run
a configure script.  Adds a small script that does a similar
job by probing a few compiler options and then outputs makefile
lines to set variables if they succeed.  These lines are added
to the default 'htscodecs.mk' file that gets built if configure
hasn't already made one.  Adding them here means the probing
will be remembered until the next "make distclean".

The script fragment that builds the default 'config.h' checks
to see if the variables have been set, and if so adds the
appropriate 'HAVE_' lines for the feature.