Add a cF CRAM specific tag.
This is used to indicate when MD and NM were not present and should
not be regenerated during decode.

Bit 1 is set when MD shouldn't be produced and bit 2 when NM shouldn't
be.  In both cases this tag is only created when embed_ref=2 and MD
and/or NM is absent from the input data.  In this scenario we cannot
reproduce the reference from SEQ+MD and cannot therefore be certain
that the value reproduced is correct.  E.g. if the reference is
produced by consensus alone, then MD is a diff of this read vs
consensus and not this read vs the original reference used in by the
aligner.

The cF tag is automatically stripped out during decode, but only with
this version of htslib and above.  Older versions will just emit a
private-space aux tag (which hopefully is harmless except for the
unlikely event of it clashing with another private name-space tool).