Fix curl type warning on Windows with gcc 12.1
curl_easy_setopt(handle, CURLOPT_POSTFIELDSIZE, val) takes a long,
but was being passed a size_t.  This works on Linux where they're
the same size for both 64 and 32-bit platforms, but not on 64-bit
Windows which has 32-bit longs and 64-bit size_t.

Casting the value to long should be fine.  The POST data is for
the completion message which should not get that big due to the limit
of 10000 parts when uploading data.