Remove limit on returned size from fai_retrieve()
This was probably a left-over from the transition to 64-bit
positions in HTSlib.  Having the limit in fai_retrieve() caused
very long references to be truncated even though programs like
`samtools faidx` should be able to support them (see issue
samtools/samtools#1660 - samtools faidx fails to retrieve large
scaffolds).

The limit is useful for legacy faidx interfaces that return
the size in an `int *`, so tests for sizes over INT_MAX
have been applied to them.