Improve embed_ref=2 when no MD:Z tags are present.
When these tags are present, the reference is inferred so it's much
the same as before.

However when not present, we previously took a first-come first-serve
approach to consensus generation as the just-in-time reference
generation was in the process_one_read loop.  As we'd already updated
the reference and delta-encoded against it, if a later read comes
along with conflicts then we couldn't correct anything.

33ff2bc incorrectly made the assumption that we were being fed a bunch
of bam records and didn't have all the data available up-front.
However this isn't true.  For the sake of threading efficiency, the
encoder buffers up all the BAMs for a container before dispatching the
cram_encode_container call, so we can do a consensus generation stage
prior to the process_one_read loop.  This improves consensus accuracy
and also simplifies the MD:Z using code too as it doesn't have to
worry about inconsistencies any more.

On 10MB worth of ONT data we had the following total file and SEQ sizes:

                   Total      SEQ only
    no_ref         205899946  74416907
    embed_ref=0    135303919  27583376  use external ref
    embed_ref=1    139709759  32008185  embed external ref, from file
    embed_ref=2+MD 139710386  32007405  embed external ref, inferred
    embed_ref=2-MD 149371457  41670856  OLD: first-come first-serve cons
    embed_ref=2-MD 139598781  31898253  NEW: proper consensus

The old embed_ref=2 needed 30% more space to delta-encode seqs,
although it was still a major win over using no_ref mode.  The new
consensus is is a significant improvement and it's now a better match
for the sequence (as we'd expect) than the external reference.
Obviously this can't be done when we have MD tags as otherwise we'd
need to store them verbatim which then takes more space up again.

On Illumina NovaSeq we also see a significant reduction to the
embedding overheads:

                   Total      SEQ only
    no_ref         256252015  87171589  48.4s
    embed_ref=0    207793068  38761828  34.7s
    embed_ref=1    219064424  50021186  35.3s
    embed_ref=2-MD 231012254  61929861  46.2s OLD
    embed_ref=2-MD 218014361  48935989  38.9s NEW

The code has also been refactored in a few places to speed it up.
Specifically with the newer consensus building strategy, we no longer
produce fake MD tags and just call the earlier algorithm, and instead
simply add directly to consensus histograms.

TODO: figure out how to enable this automatically, without penalty
when we do have the external reference available.