Add code to enable embed_ref=2 when external reference isn't found.
This choice is debatable, as it means people may get larger files than
they expect, but it also means the CRAM can be built even if
suboptimal.  A warning is given notifying the user of this action.

If a user explicitly species embed_ref=0 then this disables this
automatic code and will turn the warning back into a hard failure.

Also added more checks for handling things like unsorted data or
multi_seq_per_slice mode, neither of which are compatible with
embedding references (both the original =1 mode or =2).