Deprecate hts_filter_eval() in favour of hts_filter_eval2().
Due to hts_filter_eval() calling memset() on its res parameter,
it's not possible to pass in an allocted kstring_t in res->s
without leaking memory.  Historically it was also possible
to get away with passing in an uninitialised structure, so not
many assumptions can be made about the contents of res on entry.
In particular, it is not guaranteed that free(res->s.s) would
work.

To ensure the function is being used safely, check that the
string part of *res is NULL on entry and fail if not.
Also added a documentation note about calling hts_expr_val_free()
after hts_filter_eval().

Add hts_filter_eval2() and deprecate hts_filter_eval().  The new
function clears its `res` parameter properly, allowing it to
be called repeatedly a bit more easily than the original.