Further improve MM tag consistency checking.
If we have an MM tag with base-type specific coordinates beyond the
end of the sequence as there are too few bases of that type, then we
now detect this within bam_parse_basemod.

This was already checked within bam_next_basemod for forward reads,
but not spotted in reverse complemented ones.

Fixes #1466