Sanitize the 3-state logic for hts_expr_t, plus added some functions.
1. We have strings, numerics, and a null/unknown state.  The latter
was previously only used as a mechanism for reporting the presence of
absence of a tag, eg "[X1]" being the value of X1 if found, or null if
not.  We also had an override of is_true=1 to permit "![X1]", although
sadly this wasn't robust (and still isn't) as it has a dual meaning of
X1 not existing or X1 existing but being zero.

Our unknown state now uses NaN semantics as defined by IEE 754 for
comparisons and mathematics, but not for conditionals (see
below). This means unknown+2 is still unknown (and false).  Previously
"0/0+2" was true/-nan, it's now false/nan.

2. NaN semantics means <, >, == and != all return false with null/nan
values, even when comparing null to null.  Note this changes the
language results slightly.

INCOMPATIBILITY: Previously "[X1] != 0" meant X1 tag exists and is
zero, or X1 tag does not exist.  To avoid that second clause the man
page recommends "[X1] && [X1] != 0" to add a clause of checking the
tag for existance first.

This was illogical and almost certainly not the intended outcome.  Now
!= will be false whenever tag [X1] does not exist so the expression is
only true when the value is defined.  (The man page expression still
works, but has a redundant component.)

Similarly "![X1]" now means X1 doesn't exist, rather than the previous
interpretation of doesn't exist or is zero.

3. Fix arithmetic on non-existant aux tags.

Undefined values are now considered to be false.  They defined as
either null strings or NaN doubles, although we use the latter
ourselves.  (The former are considered the same as the previous code
used this and possibly external methods copying our old style).  Note
for compatibility with before the empty string is not false.

Previously attempting to use an undefined value gave a warning
message, so expressions like "[X0] + [X1] > 10" would spam when X0 or
X1 were absent.

Added a hts_expr_val_exists() function to simplify testing for defined
values, hts_expr_val_existsT() for defined or undef-but-true (useful
in conditionals), and hts_expr_val_undef() to set a variable to be
undefined (used when invalidating things).

4. Added an exists() function.  "[X0]" is a synonym for
X0-exists, and "![X0]" as doesn't exist, but sadly previously "![X0]"
was interpreted as "X0-doesn't-exist or X0==0".  Given this change and
for general clarity, a less ambiguous explicit exists function has
been added.

Note "exists" means has a known value or has an explicit is_true==1.
So null-but-true is still "exists".  Hence tags XX:f:nan are
considered to exist.

5. Added a default(a,b) function where "a" is returned if defined,
otherwise "b".  As previously explained the expression "[X0] + [X1] > 10"
is false whenever X0 or X1 don't exist, but using
"default([X0],0) + default([X1],0) > 10" we can use the sum of the
values present, or a single value if one is absent.

6. Added mathematical functions of sqrt, log, exp, and pow.

7. Null and boolean operations are largely unchanged, but for
clarification they work as follow, with 0/1 also being false/true and
symmetric operations.

null           == NaN (false)
null-but-true  == NaN (true)
null && x      == 0
null || 0      == 0
null || 1      == 1
!null          == 1
!!null         == 0
!null-but-true == 0

Although we're using NaN internally in order to get the arithmetic
consistent, it's not good to assume we rigidly follows all NaN
semantics.  Specifically in C NaN is considered to be true (so "NaN &&
1" is true), but for us it is false.