Make fd->embed_ref a local variable to many functions.
This field can be changed on-the-fly by cram_encode_container when it
discoveres it cannot find a reference.  It switches to embed_ref=2 in
this case.  Due to the multi-threaded nature, it is important that all
running code elsewhere is using their embed_ref parameter at the time
the container was created.

This opens up an avenue for downgrounding embed_ref again if we
require (eg a reference is absent, but subsequent ones can be found),
but for now we don't support this.