time_funcs include and dependency adjustments
Fix dependency typo in the Makefile

Remove config.h from hts_time_funcs.h as it's likely to be too
late to include it by the time that file it read.  Add config.h
include to test/test_time_funcs.c instead.

Add extra includes to hts_time_funcs.h so it stands on its own
and doesn't rely on the right headers having been included before
it is.