Improve embed_ref=2 and auto-enable it when missing external refs (PR #1449)
* Improve embed_ref=2 when no MD:Z tags are present.

When these tags are present, the reference is inferred so it's much
the same as before.

However when not present, we previously took a first-come first-serve
approach to consensus generation as the just-in-time reference
generation was in the process_one_read loop.  As we'd already updated
the reference and delta-encoded against it, if a later read comes
along with conflicts then we couldn't correct anything.

33ff2bc incorrectly made the assumption that we were being fed a bunch
of bam records and didn't have all the data available up-front.
However this isn't true.  For the sake of threading efficiency, the
encoder buffers up all the BAMs for a container before dispatching the
cram_encode_container call, so we can do a consensus generation stage
prior to the process_one_read loop.  This improves consensus accuracy
and also simplifies the MD:Z using code too as it doesn't have to
worry about inconsistencies any more.

On 10MB worth of ONT data we had the following total file and SEQ sizes:

                   Total      SEQ only
    no_ref         205899946  74416907
    embed_ref=0    135303919  27583376  use external ref
    embed_ref=1    139709759  32008185  embed external ref, from file
    embed_ref=2+MD 139710386  32007405  embed external ref, inferred
    embed_ref=2-MD 149371457  41670856  OLD: first-come first-serve cons
    embed_ref=2-MD 139598781  31898253  NEW: proper consensus

The old embed_ref=2 needed 30% more space to delta-encode seqs,
although it was still a major win over using no_ref mode.  The new
consensus is is a significant improvement and it's now a better match
for the sequence (as we'd expect) than the external reference.
Obviously this can't be done when we have MD tags as otherwise we'd
need to store them verbatim which then takes more space up again.

On Illumina NovaSeq we also see a significant reduction to the
embedding overheads:

                   Total      SEQ only
    no_ref         256252015  87171589  48.4s
    embed_ref=0    207793068  38761828  34.7s
    embed_ref=1    219064424  50021186  35.3s
    embed_ref=2-MD 231012254  61929861  46.2s OLD
    embed_ref=2-MD 218014361  48935989  38.9s NEW

The code has also been refactored in a few places to speed it up.
Specifically with the newer consensus building strategy, we no longer
produce fake MD tags and just call the earlier algorithm, and instead
simply add directly to consensus histograms.

TODO: figure out how to enable this automatically, without penalty
when we do have the external reference available.

* Add code to enable embed_ref=2 when external reference isn't found,
  and when converting BAMs with no UR or M5 tags.

This choice is debatable, as it means people may get larger files than
they expect, but it also means the CRAM can be built even if
suboptimal.  A warning is given notifying the user of this action.

If a user explicitly species embed_ref=0 then this disables this
automatic code and will turn the warning back into a hard failure.

Also added more checks for handling things like unsorted data or
multi_seq_per_slice mode, neither of which are compatible with
embedding references (both the original =1 mode or =2).

* Work around embed_ref=2 failures when MD is inconsistent.

It's too harsh to simply bail out with invalid data.  Instead we just
nullify the portion of reference computed with this record and drop
back to the consensus method.

Also replaced hts_log_warning messages with hts_log_info, as typically
we either get no messages or a huge proportion, as MD being wrong
quite often means some large scale reprocessing happened.

Tested using soft-clip adjusted covid19 samples where almost every MD
tag was incorrect.

* Make fd->embed_ref a local variable to many functions.

This field can be changed on-the-fly by cram_encode_container when it
discovers it cannot find a reference.  It switches to embed_ref=2 in
this case.  Due to the multi-threaded nature, it is important that all
running code elsewhere is using their embed_ref parameter at the time
the container was created.

This opens up an avenue for downgrading embed_ref again if we
require (eg a reference is absent, but subsequent ones can be found),
but for now we don't support this.