Adjust new variant type interfaces
* Remove `mode` from bcf_has_variant_type() interface, and add
  a special case for `VCF_REF`

Individual alleles only have a single variant type, so the only
useful mode is the overlap one (bitwise-and).  The exception
is VCF_REF, which is encoded as 0, so has to be tested for by
equality.

* Put `bcf_match_` prefix on enumerated values, to avoid name
  clashes

* Make `bitmask` unsigned for more predictable bitwise operations
  (the return value still has to be signed, though).

* Return -1 if bcf_set_variant_types() fails, of if the requested
  allele is not valid.

As callers using the legacy API won't be checking for a -1 return,
these unfortunately need to be made to call exit(1) on failure.
This is however an improvement on what would have happened under
the same conditions before, which would most likely have been a
NULL pointer dereference.

* Add a bcf_variant_length() function, to more easily access the
  rec->d.var[].n field.

* Be more specific on specifying the mask used to restrict the
  types the old functions return, in case more are added later.

* Improve documentation in the header.