Permit fasta_format to auto-switch to fastq_format.
This may be useful where hts_detect_format has reported text_format as
the lines are too long to observe the second sequence.  This change
permits us to override text_format to fasta_format and let the
fastq_parse1 function automatically select from fasta to fastq on the
first line.