Fix "make check" on MinGW.
If we haven't previously done a "make" or "make all", the .dll file
won't have been created.  The test/with-shlib.sh script will then fail
as the "cp -p ../../hts-*.dll ." command fails.  We no longer even
attempt to execute this test when plugins don't exist (which they
won't on Windows unless we've installed the dlfcn package).

Note the analogous non-mingw commands are "ln -s ../../libhts.so.* ."
and similar, which don't fail in the $? sense, but do fail with
regards to produced a bogus target containing a wild-card pattern.

This implies the error checking in this script is broken for all
platforms bar MinGW, but given the very limited use for this one test
case it's an irrelevance.