Fix "make check" on MinGW / MacOS. (PR #1496)
Fixes for when "make check" is run without first doing "make".

If we haven't previously done a "make" or "make all", the .dll file
won't have been created.  The test/with-shlib.sh script will then fail
as the "cp -p ../../hts-*.dll ." command fails.  The analogous non-mingw
commands are "ln -s ../../libhts.so.* ." and similar, which don't fail in
the $? sense, but do fail with regards to produced a bogus target
containing a wild-card pattern.

On MacOS, building with `--enable-plugins` has no dependency
between the plugins and libhts.dylib.  That also caused a test failure
because test/plugins-dlhts expects to find a dylib file.

- We no longer attempt to execute this test when the external plugins
  don't exist.

- "make plugins" now explicitly builds the shared library.

- "make check" and "make test" now has "all" as a dependency.
  While strictly not everything is a dependency of running the tests,
  ensuring all the code compiles can be viewed as one of the tests to
  perform (albeit only executed the first time due to makefile
  dependency checking rules).