Make tabix support CSI indices with large positions.
This already worked for SAM and VCF where the SQ and Contig lines
indicate the maximum length of a reference sequence.  However for BED
files this was left as zero, which had the effect of fighting against
the user by decreasing n_lvls as we increase min_shift.

When unknown, max_ref_len is now an arbitrary large size (100G), but
this may produce more levels than are strictly necessary, although
this doesn't appear to have negative consequences.

Also fixed the misleading error message about CSI being unable to
index data.  This was perhaps intended to be for mis-specified VCF
data where a contig was listed as small but the records were at larger
offsets, however it simply lead me up the garden path by categorically
stating CSI cannot store such large values.