A minor speed improvement to VCF parsing by 5-9%
Tested on 1000 genomes data (see https://github.com/brentp/vcf-bench)
this gives on 3 separate trials:

OLD
clang13 -O2 31.60 31.61 31.56
gcc12   -O2 29.32 29.36 29.33
gcc12   -O3 30.90 30.87 30.84

NEW
clang13 -O2 29.30 29.30 29.62
gcc12   -O2 27.89 27.96 27.92
gcc12   -O3 28.25 28.19 28.20
gcc7.5  -O2

The profile of NEW2 shows:
    51.41%  read     libhts.so            [.] vcf_parse_format
    21.31%  read     libhts.so            [.] bgzf_getline
     9.50%  read     libhts.so            [.] bcf_enc_vint
     9.42%  read     libdeflate.so.0      [.] deflate_decompress_default
     3.22%  read     libdeflate.so.0      [.] crc32_pclmul_avx

So that's 7.3, 4.8 and 8.6% faster for the 3 compiler tests, or about
double that for this specific function (given it's ~50% of the total
CPU).

As a bonus, it's a little less convoluted in the code too.