Check for invalid BC tags in fastq output.
I don't know how best to deal with such situations, so rather than a
hard error we treat the barcode component the same as if no BC tag was
stored - with "0".  It also whinges, and you'll probably get this
message many times, but take that as "encouragement" to fix the input
data!

Also fixed the barcode to be uppercase, incase of e.g. "ac+gt".  I'm
not sure if it's required, but we may as well.

Fixes samtools/samtools#1728