Fix ref fix from c91804c
We have too many similarly sounding reference filenames.
While the previous fixed worked when MMAP was in use, when it's not
we end up setting fd->refs->fn in cram_populate_ref, which causes
samtools test failures on Windows.

With hindsight, the fix was wrong as fd->ref_fn is the actual filename
we specified with view -T, while fd->refs->fn is the current filename
loaded (which is the same thing *unless* we're using a local cache and
no mmap in which case it's also updated to point to the filename
associated with the open file descriptor).

Definitely an "I wouldn't start from here" problem. (Sorry!)