Add CRAM SQ/M5 header checking when specifying a fasta file.
Given the possibility of creating a CRAM which cannot be decoded again
due to specifying a fasta file that differs to the M5 tags specified
in the header, we're now more militant in enforcing these match.

The references are checked on-the-fly as they're used.  For name
sorted data this may have a significant initial cost (the same that we
see when converting from BAM to CRAM without M5 tags in the BAM
header), but for position streaming the cost is only for chromosomes
currently in use.

This means writing a small CRAM file covering a region from just one
chromosome will only incur the CPU cost of validating that one
chromosome, as we're not trying to check the header tags are valid (we
take them at their word) and are instead trying to ensure the fasta
sequence we've been given matches the supplied headers.

[SQUASH_ME: This almost completely removes the previous commit, but
it's done as a separate commit for ease of review.]